feat(*): move client code to client folder
