+++ src/client/index.ts
@@ -1,0 +1,1 @@
+export { BridgeToNative } from './bridge-to-native';
