--- src/index.ts
@@ -1,2 +1,0 @@
-export { BridgeToNative } from './bridge-to-native';
-export { NativeParams, Theme, Environment, NativeFeatureKey, PdfType } from './types';
